@@ -1,42 +1,48 @@
 name: notas
 description: An application to save notes on your phone.
-author: Ana Laura Caballero Pérez
-email: anacaballero1098@gmail.com
+author: Alejandro Gómez Rivas
+email: alegomezrivas92@gmail.com
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
   # Service locator
-  get_it: ^3.1.0
+  get_it: ^7.2.0
   # Provider for handle the state
-  provider: ^4.3.2+2
+  provider: ^6.0.0
   # Value equality
-  equatable: ^1.0.1
+  equatable: ^2.0.3
   # Functional programming thingies
-  dartz: ^0.8.9
+  dartz: ^0.10.0
   # Local cache
-  hive: ^1.4.4
+  hive: ^2.0.4
   # For OS-specific directory paths
-  path_provider: ^1.6.28
+  path_provider: ^2.0.3
   # Launcher app icon
-  flutter_launcher_icons: ^0.8.1
+  flutter_launcher_icons: ^0.9.2
   # Notifications
-  awesome_notifications: ^0.0.5+5
+  awesome_notifications: ^0.0.6+10
+  # Lint
+  lint: ^1.6.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  hive_generator: ^0.8.2
-  build_runner:
+  hive_generator: ^1.1.0
+  build_runner: ^2.1.2
+  test: ^1.16.8
 
 flutter_icons:
   android: "launcher_icon"
   ios: true
   image_path: "assets/icon.png"
+  adaptive_icon_background: "assets/background.png" 
+  adaptive_icon_foreground: "assets/foreground.png"
+
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
