@@ -5,10 +5,10 @@ part 'note.g.dart';
 @HiveType(typeId: 1)
 class Note {
   @HiveField(0)
-  final String title;
+  final String? title;
 
   @HiveField(1)
-  final String content;
+  final String? content;
 
   Note({this.title, this.content});
 
