@@ -1,6 +1,5 @@
 import 'package:equatable/equatable.dart';
 import 'package:notas/core/usecase/usecase.dart';
-import 'package:meta/meta.dart';
 import 'package:dartz/dartz.dart';
 import 'package:notas/core/error/failures.dart';
 import 'package:notas/features/notes/domain/entities/note.dart';
@@ -9,18 +8,18 @@ import 'package:notas/features/notes/domain/repositories/note_repository_abstrac
 class AddNote implements UseCase<void, Params> {
   AddNote(this.repository);
 
-  final NoteRepository repository;
+  final NoteRepository? repository;
 
   @override
   Future<Either<Failure, void>> call(Params params) async {
-    return await repository.addNote(params.note);
+    return await repository!.addNote(params.note);
   }
 }
 
 class Params extends Equatable {
   final Note note;
 
-  Params({@required this.note});
+  Params({required this.note});
 
   @override
   List<Object> get props => [note];
