@@ -1,5 +1,4 @@
 import 'package:dartz/dartz.dart';
-import 'package:meta/meta.dart';
 import 'package:equatable/equatable.dart';
 import 'package:notas/core/usecase/usecase.dart';
 import 'package:notas/core/error/failures.dart';
@@ -9,11 +8,11 @@ import 'package:notas/features/notes/domain/repositories/note_repository_abstrac
 class EditNote implements UseCase<void, Params> {
   EditNote(this.repository);
 
-  final NoteRepository repository;
+  final NoteRepository? repository;
 
   @override
   Future<Either<Failure, void>> call(Params params) async {
-    return await repository.editNote(params.index, params.note);
+    return await repository!.editNote(params.index, params.note);
   }
 }
 
@@ -21,7 +20,7 @@ class Params extends Equatable {
   final int index;
   final Note note;
 
-  Params({@required this.index, @required this.note});
+  Params({required this.index, required this.note});
 
   @override
   List<Object> get props => [index, note];
