@@ -1,7 +1,7 @@
 import 'package:equatable/equatable.dart';
 
 abstract class Failure extends Equatable {
-  Failure([List propierties = const <dynamic>[]]) : super();
+  const Failure() : super();
 }
 
 // General failures
