@@ -17,8 +17,8 @@ class NoteAdapter extends TypeAdapter<Note> {
       for (int i = 0; i < numOfFields; i++) reader.readByte(): reader.read(),
     };
     return Note(
-      title: fields[0] as String,
-      content: fields[1] as String,
+      title: fields[0] as String?,
+      content: fields[1] as String?,
     );
   }
 
