@@ -4,7 +4,7 @@ import 'package:notas/features/notes/domain/entities/note.dart';
 
 abstract class NoteRepository {
   Future<Either<Failure, void>> addNote(Note note);
-  Future<Either<Failure, void>> editNote(int index, Note note);
+  Future<Either<Failure, void>> editNote(int? index, Note note);
   Future<Either<Failure, void>> deleteNote(int index);
   Future<Either<Failure, List<Note>>> getAllNotes();
 }
