@@ -50,7 +50,7 @@ class _MyAppState extends State<MyApp> {
       builder: (context, child) {
         return MediaQuery(
           data: MediaQuery.of(context).copyWith(textScaleFactor: 1.0),
-          child: child,
+          child: child!,
         );
       },
       home: NotePage(),
